[client] ProductListコンポーネントの作成
